move json.lua to lower case and more error checking in record_message
